Split symbol list off to its own reusable widget
